requirements.txt
Flask==0.12
PyGitHub==1.32